Reduced meat cost
Now I know the argument against doing this is that we should leave it until we fix the mob spawning/culling plugin so the original numbers will be possible. However, people will not be able to smoothly test the recipes in their current state and the only complaints will be for the meat. For the sake of good testing, I have reduced the cost according to feedback from a few players working on the recipes.