Final XP Changes
After feedback from people farming for the grandmaster recipe with enchanted tools, the wood was taking up roughly 90% of the time to farm. Any person (even without having huge automatic farms which would even further cry out for this change) making xp has been saying that the wood is not only highly disproportionate, but a majority of the time-sink in relation to the other materials combined.

These changes is essentially also making XP cheaper again, which handles with the repair costs to tools/armour being introduced soon (tm). The repair costs will be 2.5% roughly of the enchantment cost on each item, and after running maths on multiple people using enchanted tools it becomes a very big problem for any place to be able to sustain their tools.

From the current CivTemp values (before the changes in the last pull request which doubled the material costs) these are the changes:

- Quarter aether cost of each recipe.
- Half emerald output of each recipe.
- Reduced production time accordingly as more runs are required.
- Adding repair recipes at half of the current proposed cost (for tools/armour). *these changes are part of a different config*
- Cutting wood costs:
  - Apprentice - Half (64 Logs - 32 Logs)
  - Adept - Quarter (5stacks - 2.5stacks)
  - Grandmaster - Eigth (25stacks - 3stacks)