Updated Fossils according to feedback.
People were mentioning that they were getting too many low-tier drops that were clogging up their inventory, and it was generally a hassle as it is tedious to get a load of stone tools and leather armor that should probably be burnt.

In this version, there is a more defined 'you lose' event, as it will drop a piece of cobblestone 99 times out of 100 (pretty much). Then 1 in 100 times you get rewarded with a mid-tier or above drop, which are actually things people wouldn't mind getting! This results in fossils being much more like scratchcards - they aren't a hassle when you lose but nearly all of the time you do lose, but there are chances to strike it lucky.

There has also been criticism on the possible outcome of a dangerous and competent raider getting their hands on a Sharpness V book on day 1 (or early game) and flattening the entire server. After a little bit of working out I can show the actual likely-hood of the situation:

Let's suppose there are 15 bad players on launch day that are both very good at PvP or know someone that is (and willing to donate to them), and also looking to raid towns and generally be a threat. Let's also suppose that each one of them mines out 15,000 blocks in the first couple days (which is a very large amount on average but for these players it is possible). That's 225,000 stone mined by baddies who will gather up a lot of fossils as a result.

To get a Sharpness V book from a fossil, on average it would take 8,270,682 stone to have been mined until one should have appeared. Tying this in to our scenario, it results in a 2.7% chance that this scenario will happen. This is still counting on there actually being this many threatening players on the server, and all of them mining that amount. The event is not impossible, but it is far from likely and will barely impact on any long-term balances inside the game. In the instance that this event did occur, then bear in mind that at that stage in the game the raider will also have the same level of armour as everyone else in the game - still putting them far from a full Prot IV Sharp V load-out monster that could destroy anything he/she pleases with no risk.

So yea - it's all good. I think we can give this a try for a few days at least so that players can adjust to these types of fossils and understand their purpose a little more before taking in too much more feedback (as players are still used to the much rarer fossils that will always get something good, they need a bit of time to get used to them dropping something with no value more often).